Add back storage bag, diamond anvil and speed charm age one.
Storage bag and speed charm are only useful earlier on.  Don't need to be speeding around once at stage 5 etc.
Inventory management quickly becomes a problem early on.  Provide some kind of salve.

Diamond anvil for lulz, maybe craftable with workarounds, but unlikely.  Still a nice to have as a reward if player is able to get around other restrictions.